naukri-bot/requirements.txt
selenium
google-generative-ai
webdriver-manager
beautifulsoup4
pandas
